+++ ostd/specs/mm/page_table/cursor/tree_lemmas.rs
@@ -42,13 +42,13 @@ impl<'rcu, C: PageTableConfig> CursorContinuation<'rcu, C> {
             #![auto]
             0 <= j < self.children.len()
                 && self.children[j] is Some implies self.children[j].unwrap().subtree_satisfies(
-            self.path().push_tail(j as usize),
+            self.path().push_tail(j),
             g,
         ) by {
             self.inv_children_unroll(j);
             OwnerSubtree::lemma_map_implies(
                 self.children[j].unwrap(),
-                self.path().push_tail(j as usize),
+                self.path().push_tail(j),
                 f,
                 g,
             );
@@ -76,7 +76,7 @@ impl<'rcu, C: PageTableConfig> CursorContinuation<'rcu, C> {
                 #![auto]
                 0 <= j < NR_ENTRIES && j != idx ==> self.children[j] == cont0.children[j],
             self.children[idx] is Some ==> self.children[idx]->0.subtree_satisfies(
-                self.path().push_tail(idx as usize),
+                self.path().push_tail(idx as int),
                 g,
             ),
         ensures
@@ -86,14 +86,14 @@ impl<'rcu, C: PageTableConfig> CursorContinuation<'rcu, C> {
             #![auto]
             0 <= j < self.children.len()
                 && self.children[j] is Some implies self.children[j].unwrap().subtree_satisfies(
-            self.path().push_tail(j as usize),
+            self.path().push_tail(j),
             g,
         ) by {
             if j != idx {
                 cont0.inv_children_unroll(j);
                 OwnerSubtree::lemma_map_implies(
                     cont0.children[j].unwrap(),
-                    cont0.path().push_tail(j as usize),
+                    cont0.path().push_tail(j),
                     f,
                     g,
                 );
@@ -144,11 +144,11 @@ impl<'rcu, C: PageTableConfig> CursorOwner<'rcu, C> {
                 #![trigger cont.children[j]]
                 0 <= j < cont.children.len()
                     && cont.children[j] is Some implies cont.children[j].unwrap().subtree_satisfies(
-            cont.path().push_tail(j as usize), g) by {
+            cont.path().push_tail(j), g) by {
                 cont.inv_children_unroll(j);
                 OwnerSubtree::lemma_map_implies(
                     cont.children[j].unwrap(),
-                    cont.path().push_tail(j as usize),
+                    cont.path().push_tail(j),
                     f,
                     g,
                 );
@@ -239,13 +239,13 @@ impl<'rcu, C: PageTableConfig> CursorOwner<'rcu, C> {
             assert forall|j: int|
                 0 <= j < NR_ENTRIES
                     && #[trigger] cont.children[j] is Some implies cont.children[j].unwrap().subtree_satisfies(
-            cont.path().push_tail(j as usize), g) by {
+            cont.path().push_tail(j), g) by {
                 cont.inv_children_unroll(j);
                 PageTableOwner::tree_not_in_scope(
                     cont.children[j].unwrap(),
-                    cont.path().push_tail(j as usize),
+                    cont.path().push_tail(j),
                 );
-                cont.children[j].unwrap().lemma_map_implies(cont.path().push_tail(j as usize), nsp, g);
+                cont.children[j].unwrap().lemma_map_implies(cont.path().push_tail(j), nsp, g);
             };
         };
     }
